test(mempool_p2p): test tx received from p2p reaches mempool
